// src/types/PROFILE/enum/EnumTypeAccount.ts
export enum EnumTypeAccount {
 MUSICIAN = "musician",
 TEAM = "team",
 INSTITUTION = "institution",
 MUSICIAN_SCHOOL = "musicianSchool",
 REHEARSAL_BASE = "rehearsalBase",
 MUSIC_SHOP = "musicShop",
 MUSICAL_WORKSHOP = "musicalWorkshop",
 SOUND_PRODUCER = "soundProducer",
 RECORDING_STUDIO = "recordingStudio",
}

//Музыкант
//Команда
//Учреждение
//МузыкантШкола
//Репетиционная база
//Звукозаписывающая студия
//Музыкальный магазин
//МузыкальнаяМастерская
//Саундпродюсер